velpari_youtube/requirements.txt
elevenlabs==0.2.26
openai==1.3.0
moviepy==1.0.3
requests==2.31.0
pyyaml==6.0.1
pytest==7.4.3
python-dotenv==1.0.0
tqdm==4.66.1
httpx==0.25.2 